Relocate SelectItem
