+++ asreview/webapp/src/ProjectComponents/SetupComponents/ModelComponents/ModelForm.js
@@ -16,7 +16,8 @@ import {
 import { styled } from "@mui/material/styles";
 
 import { InlineErrorHandler } from "../../../Components";
-import { InfoCard, SelectItem } from "../../SetupComponents";
+import { SelectItem } from "../../../ProjectComponents";
+import { InfoCard } from "../../SetupComponents";
 import { ProjectAPI } from "../../../api/index.js";
 import { defaultAlgorithms, mapStateToProps } from "../../../globals.js";
