+++ asreview/webapp/src/ProjectComponents/SetupComponents/index.js
@@ -1,5 +1,4 @@
 export { default as FinishSetup } from "./FinishSetup";
 export { default as InfoCard } from "./InfoCard";
 export { default as SavingStateBox } from "./SavingStateBox";
-export { default as SelectItem } from "./SelectItem";
 export { default as SetupDialog } from "./SetupDialog";
