+++ asreview/webapp/src/ProjectComponents/index.js
@@ -4,3 +4,4 @@ export { default as ProjectImportDialog } from "./ProjectImportDialog";
 export { default as ProjectInfoForm } from "./ProjectInfoForm";
 export { default as ProjectModeSelect } from "./ProjectModeSelect";
 export { default as ProjectPage } from "./ProjectPage";
+export { default as SelectItem } from "./SelectItem";
